LayoutTests/crypto/subtle/aes-cbc-import-key-wrap-jwk-rsa-key-private-expected.txt

Test wrapping a JWK RSA private key with AES-CBC using an imported key

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS exportedKey.d is jwkKey.d
PASS exportedKeyPhiFormat.p is jwkKeyPhiFormat.p
PASS exportedKeyPhiFormat.q is jwkKeyPhiFormat.q
PASS bytesToHexString(wrappedKey) is expectWrappedKey
PASS successfullyParsed is true

TEST COMPLETE
